Change values from int to long for hate
Implement some dragon valley raids ai
Thanks to l2j

Implement vangara script zone for future usage

Core:

compile and copy paste core.jar

Data:

Delete scripts folder and paste from this rev

+++ java/l2r/gameserver/ai/L2DoorAI.java
@@ -95,7 +95,7 @@ public class L2DoorAI extends L2CharacterAI
 	}
 	
 	@Override
-	protected void onEvtAggression(L2Character target, int aggro)
+	protected void onEvtAggression(L2Character target, long aggro)
 	{
 	}
